chore: move PatchPanel to components
